smooth light levels

+++ src/components.rs
@@ -25,7 +25,7 @@ pub struct Player {}
 pub struct Viewshed {
     pub visible_tiles: Vec<Position>, // positions relative to algorithm that are visible
     pub light_levels: Vec<Option<f32>>, // light levels for visible tiles
-    pub strength: usize, // effects range of shadowcasting algorithm
+    pub strength: f32, // affects how far light extends from source & light level dropoff rate from origin
     pub dirty: bool, // has game changed (player moved)?
 }
